refactor: Refactor headers
