+++ xero/pipeline/credit_notes.py
@@ -1,7 +1,7 @@
 from xero.pipeline.interface import Pipeline
 from xero.repo import get_listing
 from xero.pipeline.utils import parse_timestamp
-from xero.pipeline.headers_builder import timeframe
+from xero.pipeline.headers import timeframe
 
 pipeline = Pipeline(
     "CreditNotes",
